// src/common/enums/error.enum.ts
export enum EError {
  // Module auth, driver
  INVALID_PASSWORD = 'INVALID_PASSWORD',
  INVALID_EMAIL = 'INVALID_EMAIL',
  INVALID_PHONE_NUMBER = 'INVALID_PHONE_NUMBER',
  INVALID_AVATAR = 'INVALID_AVATAR',
  INVALID_IDENTITY_CARD = 'INVALID_IDENTITY_CARD',
  DRIVER_EXIST = 'DRIVER_EXIST',
  CANNOT_FIND_USER = 'CANNOT_FIND_USER',
  INVALID_USER_STATUS = 'INVALID_USER_STATUS',
  SIGN_UP_ERROR = 'SIGN_UP_ERROR',
  INVALID_INDETITY = 'INVALID_IDENTITY',
  INVALID_REFRESH_TOKEN = 'INVALID_REFRESH_TOKEN',
  TOKEN_IN_BLACKLIST = 'TOKEN_IN_BLACKLIST',
  DRIVER_NOT_FOUND = 'DRIVER_NOT_FOUND',
  INVALID_FULL_NAME = 'INVALID_FULL_NAME',
  INVALID_ACTIVE_AREA = 'INVALID_ACTIVE_AREA',
  INVALID_TEMPORARY_ADDRESS = 'INVALID_TEMPORARY_ADDRESS',
  UPDATE_DRIVER_REFRESH_TOKEN_ERROR = 'UPDATE_DRIVER_REFRESH_TOKEN_ERROR',
  INVALID_CODE_DRIVER_REQUEST = 'INVALID_CODE_DRIVER_REQUEST',
  INVALID_DESCRIPTION_DRIVER_REQUEST = 'INVALID_DESCRIPTION_DRIVER_REQUEST',
  INVALID_REQUEST_TYPE_ID = 'INVALID_REQUEST_TYPE_ID',
  INVALID_FILE_ID = 'INVALID_FILE_ID',
  INVALID_DRIVER_REQUEST_ID = 'INVALID_DRIVER_REQUEST_ID',

  // Module upload, file
  NO_FILE_PROVIDED = 'NO_FILE_PROVIDED',
  FILE_TOO_LARGE = 'FILE_TOO_LARGE',
  INVALID_FILE_SIZE = 'INVALID_FILE_SIZE',
  INVALID_FILE_EXTENSION = 'INVALID_FILE_EXTENSION',
  INVALID_FILE_TYPE = 'INVALID_FILE_TYPE',
  UPLOAD_FAILED = 'UPLOAD_FAILED',
  MINIO_CONNECTION_ERROR = 'MINIO_CONNECTION_ERROR',
  BUCKET_NOT_FOUND = 'BUCKET_NOT_FOUND',
  FILE_NOT_FOUND = 'FILE_NOT_FOUND',
  DELETE_FAILED = 'DELETE_FAILED'
}